+++ packages/sonata-api/src/loader.ts
@@ -1,7 +1,7 @@
 import type { init } from '@sonata-api/server'
 
 const main = async () => {
-  const entrypoint: ReturnType<typeof init> = (await (new Function('return import(process.argv[1])'))()).default
+  const entrypoint: ReturnType<typeof init> = require(process.argv[1]).default
   entrypoint.listen()
 }