+++ packages/types/src/index.ts
@@ -11,5 +11,9 @@ export * from './property.js'
 export * from './schema.js'
 export * from './security.js'
 export * from './validation.js'
-export { ObjectId } from 'mongodb'
-export type { WithId, OptionalId, } from 'mongodb'
+export {
+  ObjectId, 
+} from 'mongodb'
+export type {
+  WithId, OptionalId, 
+} from 'mongodb'