+++ packages/builtins/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sonata-api/system
 
+## 1.0.27
+
+### Patch Changes
+
+- ESM extensions
+- Updated dependencies
+  - @sonata-api/types@0.1.41
+  - @sonata-api/access-control@1.0.16
+  - @sonata-api/api@1.0.41
+  - @sonata-api/common@0.1.20
+  - @sonata-api/entrypoint@0.0.12
+  - @sonata-api/validation@1.0.15
+
 ## 1.0.26
 
 ### Patch Changes