+++ packages/http/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sonata-api/http
 
+## 1.0.20
+
+### Patch Changes
+
+- ESM extensions
+- Updated dependencies
+  - @sonata-api/types@0.1.41
+  - @sonata-api/access-control@1.0.16
+  - @sonata-api/common@0.1.20
+  - @sonata-api/validation@1.0.15
+
 ## 1.0.19
 
 ### Patch Changes