+++ packages/server/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sonata-api/server
 
+## 1.0.19
+
+### Patch Changes
+
+- ESM extensions
+- Updated dependencies
+  - @sonata-api/types@0.1.41
+  - @sonata-api/access-control@1.0.16
+  - @sonata-api/api@1.0.41
+  - @sonata-api/builtins@1.0.27
+  - @sonata-api/common@0.1.20
+  - @sonata-api/entrypoint@0.0.12
+  - @sonata-api/http@1.0.20
+  - @sonata-api/server@1.0.19
+
 ## 1.0.18
 
 ### Patch Changes